+++ js/map.js
@@ -1,8 +1,7 @@
 import {removeDisabledStatePage} from './toggle-state-page.js';
+import {SIMILAR_AD_COUNT, similarAds} from './data.js';
 
 const validationForm = document.querySelector('.ad-form');
-
-const reset = validationForm.querySelector('.ad-form__reset');
 
 const addressValueInput = validationForm.querySelector('#address');
 
@@ -14,24 +13,20 @@ const CenterTokyo = {
 };
 
 const mainPin = {
-  URL: 'img/main-pin.svg',
-  SIZE: [40, 40],
-  ANCHOR: [20, 40],
+  url: 'img/main-pin.svg',
+  size: [40, 40],
+  anchor: [20, 40],
 };
 
 const pins = {
-  URL: 'img/pin.svg',
-  SIZE: [52, 52],
-  ANCHOR: [26, 52],
-};
-
-const onMapLoad = function () {
-  removeDisabledStatePage();
+  url: 'img/pin.svg',
+  size: [52, 52],
+  anchor: [26, 52],
 };
 
 const map = L.map('card-canvas')
   .on('load', () => {
-    onMapLoad();
+    removeDisabledStatePage();
   })
   .setView({
     lat: CenterTokyo.LAT,
@@ -47,9 +42,9 @@ L.tileLayer(
 ).addTo(map);
 
 const mainPinIcon = L.icon({
-  iconUrl: mainPin.URL,
-  iconSize: mainPin.SIZE,
-  iconAnchor: mainPin.ANCHOR,
+  iconUrl: mainPin.url,
+  iconSize: mainPin.size,
+  iconAnchor: mainPin.anchor,
 });
 
 const mainPinMarker = L.marker(
@@ -70,9 +65,9 @@ mainPinMarker.on('move', (evt) => {
 });
 
 const pinIcon = L.icon({
-  iconUrl: pins.URL,
-  iconSize: pins.SIZE,
-  iconAnchor: pins.ANCHOR,
+  iconUrl: pins.url,
+  iconSize: pins.size,
+  iconAnchor: pins.anchor,
 });
 
 const addPinsMarker = function (ads) {
@@ -98,7 +93,15 @@ const setMainPin = function () {
     lat: CenterTokyo.LAT,
     lng: CenterTokyo.LNG,
   }, ZOOM_MAP);
-
 };
 
-export {setMainPin, onMapLoad, CenterTokyo, validationForm, addPinsMarker, map, reset};
+const getDataPins = function (adverts) {
+  addPinsMarker(adverts.slice(0, SIMILAR_AD_COUNT));
+};
+
+function loadMap () {
+  getDataPins(similarAds);
+  setMainPin(CenterTokyo);
+}
+
+export {setMainPin, loadMap, CenterTokyo};
